requirements.txt
django==4.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
pyjwt==2.4.0